% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_salaries.R
\docType{data}
\name{ds_salaries}
\alias{ds_salaries}
\title{Data science salaries dataset}
\format{
A data frame with 3755 rows and 11 variables:
\describe{
\item{work_year}{Year of work}
\item{experience_level}{Experience level}
\item{employment_type}{Employment type}
\item{job_title}{Job title}
\item{salary}{Salary in original currency}
\item{salary_currency}{Currency of salary}
\item{salary_in_usd}{Salary converted in USD}
\item{employee_residence}{Employee residence}
\item{remote_ratio}{Remote work ratio}
\item{company_location}{Company location}
\item{company_size}{Company size}
}
}
\source{
\url{https://www.kaggle.com/datasets/saurabhshahane/data-science-jobs-salaries}
}
\usage{
ds_salaries
}
\description{
A dataset containing salaries in data science jobs.
}
\keyword{datasets}
