% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_salaries.R
\name{filter_salaries}
\alias{filter_salaries}
\title{Filter salaries by experience level and company size}
\usage{
filter_salaries(data, experience_level, company_size)
}
\arguments{
\item{data}{A data frame}

\item{experience_level}{A character string}

\item{company_size}{A character string}
}
\value{
A filtered data frame
}
\description{
Filter salaries by experience level and company size
}
\examples{
filter_salaries(ds_salaries, "SE", "M")
}
