% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_salary_by_experience.R
\name{mean_salary_by_experience}
\alias{mean_salary_by_experience}
\title{Mean salary by experience level}
\usage{
mean_salary_by_experience(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Mean salary by experience level
}
\examples{
mean_salary_by_experience(ds_salaries)
}
