% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_salary_by_experience.R
\name{plot_mean_salary_by_experience}
\alias{plot_mean_salary_by_experience}
\title{Plot mean salary by experience level}
\usage{
plot_mean_salary_by_experience(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A ggplot object
}
\description{
Plot mean salary by experience level
}
\examples{
plot_mean_salary_by_experience(ds_salaries)
}
